#174 Refactoring om memory sub system. Broke it up in multiple files
